Docker Deps & Ease of use
Update the docker deps to 8.3 and add some 'ease of use' to
the docker file, vscode in sudo and some basic packages.

These tools make it easier for debugging in the container in vscode.

By not tying the DB exposed to 3306 we can also run multiple containers on the one host.